Yarn more than one workspace problem
When the workspaces option is set as a glob, it appears that Yarn looks
into sub node_modules an picks workspaces that it should not. For example,
in this repository, Yarn complains that ramda is there more than once
because it exists in mutliple node_modules location.

To workaround the issue, one can specify each package independtly
right now.